fix: fixed visual issues

+++ Game Project/Assets/Canvas/Scripts/PlayerBalance.cs
@@ -1,7 +1,7 @@
 public class PlayerBalance
 {
 
-    public int Coins { get; set; } = 0;
+    public float Coins { get; set; } = 0;
     public int Points { get; set; } = 0;
     public PlayerBalance()
     {
@@ -10,7 +10,7 @@ public class PlayerBalance
 
     public void AddCoins(int coins) => this.Coins += coins;
     public void AddPoints(int points) => this.Points += points;
-    public void SubtractCoins(int coins) => this.Coins -= coins;
+    public void SubtractCoins(float coins) => this.Coins -= coins;
     public void SubtractPoints(int points) => this.Points -= points;
 
 }
