+++ Game Project/Assets/Prefabs/UI/CardScript.cs
@@ -2,7 +2,7 @@ using UnityEngine;
 
 public class CardScript : MonoBehaviour
 {
-    public int priceValue;
+    public float priceValue;
     public int level = 1;
     public float value = 10;
     public float upgradeValue = 5;
